MH-13444, Insecure Series Creation
This patch prevents users from being able to overwrite series metadata
by default when ingesting event data. Users can still enable this, if
they need this functionality.

In addition to this, Opencast will not any longer add anonymous access
to series if they are created as part of the event ingest.

*Note that the behavior of both features is questionable and we may want
to remove this entirely in the future.*