MH-13455, Opencast Plug-in Features
This patch extracts some core features into separate Karaf features.
These features are installed by default, but not started by default.
This means that they can be easily enabled but will not launch
regardless of users needing those features or not.

The idea behind this is to extract modules used by just a few users to
minimize the code which is running by default, needing less resources,
booting up faster and also minimizing the risk for security
vulnerabilities.

This patch extracts the Moodle and Sakai user provider as an example.
More modules may follow.

Long-term, it may make sense to still boot all features in the
development distribution to ensure they are not easily broken without
someone noticing. That will be easier once KARAF-6207 is fixed:

- https://github.com/apache/karaf/pull/791