MH-13462, Prevent Being Started By Root
This patch changes the configuration to prevent Opencast from being
started by a root user unless the configuration is changed.

We have seen several times in the past that someone launched Opencast
manually as root user on e.g. a RPM or deb install, causing a number of
artifacts to be written with root privileges so that for later usage,
access rights for all these files needed to be modified again.