MH-13482 Reset the job load to 0 if the job load cache is empty but the local system load is not zero
This should not happen if the logic elsewhere is correct, so this is a failsafe rather than
expected condition (hence the WARN logging).

The consequences of a worker node having a long-term incorrect job load are that the node will either
be under-used or over-used, so it's best to correct this if known to be wrong.