MH-13503: Initial work dispatching to node with the lowest utilization.
The initial implementation of the job load code made the assumption that the nodes were more or less equal in terms of
number of cores.  This is obviously a bad assumption, and led to workers with a smaller number of cores being tapped
far harder than a larger worker because the worker's maximum load was not considered.

This commit replaces that behaviour, instead considering the current vs maximum load at dispatch time for each node.