Allow different sources for different resolutions with influx
This patch changes the handling of statistics configurations such that
for each provider type, there can be a different configuration format.
This makes the mechanism more flexible and reflects the actual variety
of options that might come with different sources. This way, we also get
rid of the "source" property of the configuration which used to be a
compisite string. This didn't make too much sense since we were using
JSON already.

The new configuration format allows to specify different sources for
one influx provider, so that one can use specific measuemrents for
certain data resolutions. In practice you might want to use shorter
retentions for finer-grained measurements. This change enables you to do
so.