MH-13503: Altering dispatch logic to take maxload and load factor into account.
The goal here is to dispatch to the lowest load factor, with the node's maximum load being the tie breaker.  This
ensures that nodes with a large amount of remaining capacity are dispatched to first, rather than attempting to
keep a uniform (numerically) even load in terms of absolute load consumed.