MH-13535 Add basic CSV export for statistics
This patch adds "Download" buttons to the statistics graphs in the
Admin UI. Pressing these buttons will trigger a CSV file download. The
content of the CSV file depends on the current settings of the graph
(e.g. from, to, granularity). Basically, the data seen in the graph is
exported to the CSV.

For series statistics data, the behavior is configurable: with the
default settings, you just get the data you see (just like for episode
and organization). With an advanced configuration, you get the data of
all events of the series, instead. To enable this, you have to configure
which statistics providers should be used to retrieve the events'
statistics.