MH-13542 Display global notifications as overlay
In Opencast's admin UI, notifications are used across various places.
This patch changes the look and feel of "global" notifications. By
"global" notifications, we mean those notifications, which are not
displayed inside a modal or wizard.

With this patch, global notifications appear as boxes in the top right
corner of the admin UI. This has some advantages:

- They are always visible, regardless of the scroll position within the
  page
- They don't take space and hence, do not move all subsequent html
  elements