MH-13554, Theodul Zoom
One suggestion from the final presentation of the player usability study
at Osnabrück University was to change the zoom function bound to the
mouse wheel since trying to scroll down the page would often engage the
zoom functionality.

A suggested solution was to combine the mouse wheel event with a
modifier key like shift which is exactly what this patch does. So, in
addition to the `plus` and `minus` keys and the menu, `shift + wheel`
will engage the zoom functionality.