MH-13557, Login Response for JavaScript
All JavaScript code letting user log-in to Opencast are using some kind
of ugly hack right now. For example, the engage interface is evaluating
if the response HTML contains a particular string.

For JavaScript, it would be much easier, better and more reliable if we
could be a proper HTTP status and JSON response.