MH-13564, Media Module Configuration
This patch moves the media module configuration from the organization
configuration to the user interface configuration service to provide a
better structure for those components.

Instead of using JSON directly, I opted for using YAML for the
configuration file since that allows us to add proper comments to the
configuration file.