MH-13584, Jettison Dependency Management
Jettison is defined as dependency in multiple modules including the
version which unfortunately differs from the version cxf depends on
which results in Opencast's distributions containing multiple versions
of this library.

This patch centralizes the dependency management of Jettison and
upgrades the library to the same version Karaf already contains.