MH-13522 Add option to configure state mappings for workflows (#861)
This patch adds the possibility to define so-called state mappings for
workflows. State mappings tell what to display in the admin UI when a
workflow is in one of the mapped workflow states. For example, you can
configure that the admin UI should display "Retracting" instead of
"Running" when the retract workflow is active. When no state mappings
are given, the default mappings are used (exactly as before). This is
mainly an usability improvement.

No migration or index rebuild is needed to make use of the new mappings.
Just extend your workflow definitions accordingly.