MH-13602, Update jackson-databind to fix CVE-2019-12086
- [CVE-2019-12086](https://nvd.nist.gov/vuln/detail/CVE-2019-12086)
- moderate severity
- Vulnerable versions: >= 2.0.0, < 2.9.9
- Patched version: 2.9.9

A Polymorphic Typing issue was discovered in FasterXML jackson-databind
2.x before 2.9.9. When Default Typing is enabled (either globally or for
a specific property) for an externally exposed JSON endpoint, the
service has the mysql-connector-java jar (8.0.14 or earlier) in the
classpath, and an attacker can host a crafted MySQL server reachable by
the victim, an attacker can send a crafted JSON message that allows them
to read arbitrary local files on the server. This occurs because of
missing com.mysql.cj.jdbc.admin.MiniAdmin validation.

---

Shouldn't be a high risk for Opencast but better safe than sorry.