MH-13612, Tag elements retrieved from asset manager
By now, it seems to be the general assumptions that elements stored in
and then retrieved from the asset manager have the tag `archive`.

While this is often the case as the `archive` tag is widely used to
flag elements for archival, this is by no means guaranteed and this
misconception has lead to multiple issues over the last years.

Multiple workarounds and partial fixes have been added for this, e.g.:

  https://github.com/opencast/opencast/blob/49fe433b42d6dd1ec59c7364d23ff1060f19040a/etc/workflows/fast.xml#L81-L83

…but this can still easily happen again.

This patch finally ensures that all elements which are retrieved from
the asset manager as part of a media package are guaranteed to have the
`archive` tag, finally bringing an end to workarounds and uncertainty.