MH-13648, Asset Manager Concurrecy Issue
The asset managers filesystem storage backend may fail to create
directories if it tries to create the directory multiple time in
parallel.

The reason for this is an unnecessary, non-synchronized check if the
directory already exists before trying the create the directory
(`mkdirs()` already includes this check) and the incorrect assumption
that `mkdirs()` returning false means that the directory is not created.