MH-13650, Scheduler Migration Performance
Depending on the amount of asset manager properties in the system, the
migration of scheduled events between Opencast 6 and Opencast 7 can be
very slow.

In our tests on SWITCH's test cluster (approx. 1M properties in the
database) migration for a single event took 2-3 seconds. That makes a
full migration take forever.

This patch brings this time down to approx 0.01 seconds on the same
system by using the asset managers native queries which are a lot
faster.