MH-13660, Update Development Process Documentation
The development process documentation has grown old at some places and
does not completely reflect current practice any longer. Since the
document was written, the version numbers have changed, we do not use
Bitbucket any longer and we usually do not cut beta releases or release
candidates.

Additionally, to updating that, this patch includes a Markdown extension
to write inline Graphviz code for rendering vector graphics. We've
discussed adding something like this a while ago and this seemed a good
opportunity to get it done.