MH-13670, Shell information for developer distribution
Launching the developer distribution of Opencast, `log:tail` is invoked
automatically to make testing easier. This might confuse first-time
users of the development distribution since they need to know that they
can hit `ctrl-c` to leave the command.

This patch adds a comment about how to terminate a tool to all
distributions and a comment about automatically launching `log:tail` to
the development distribution.