MH-13672, Editor Maximum Height
The player used in the video editor has no maximum height which max
cause the video to take more than the full height of the screen in
certain situations (e.g. high-res portrait mode smartphone recording).
This makes the editor extremely hard to use in these situations.

A simple solution is to enforce a maximum height, the player max not
exceed.