MH-13697, Workflow Index Rebuild Memory
For an index rebuild, the workflow service will load all workflows into
memory at once. Depending on the amount of workflows in the system, this
can cause a huge, sudden spike in memory usage which may even cause the
whole process to fail. E.g. on one larger system we managed to kill the
process on a machine with 64GB memoty…

This patch introduces pagination for the workflows. This means that only
a limited number of workflows are loaded from the database and processed
at the same time.