MH-13516 add single step event deletion (#862)
This patch changes the event deletion via admin UI such that you don't
have to retract events before you are able to delete them. While that
was the case in the past, now, Opencast itself takes care of retracting
the events before deleting them if, necessary. This also applies for
bulk deletion.

This is achieved by first checking if the event to delete still has
publications and if so, running a retract workflow and waiting for its
completion. Once, a retract workflow is completed, the associated event
is deleted.

The criteria to tell if an event needs to be retracted is equal to the
one that was used in the admin UI frontend before.

The id of the workflow to use for event retraction is configurable.