MH-13714, Hide Column `Stop` By Default
As discussed a few times already, the column containing the technical
end date of the events table is not very helpful for most people since
it exists only for parts of the recordings.

This patch thus updates the default settings of the admin interface,
hiding the column initially. Users are still able to configure the
columns like before and can still include the column if they want to.

This patch has no effect if a custom setting is already stored in the
user's local storage.