Remove State Mapping Importing
This patch removes the state mapping “Importing” from the default
workflow since it can be pretty confusing if it is used to directly
publish material and the workflow state is forever shown as “Importing”
only to jump to finish right at the end.

While “Importing” is nice when the editor is used, the more general
default mapping “Running” fits better when both use-cases are
considered.