MH-13706 fixes rendered event date values in the events overview of the admin UI
With the PR #1049 the event start date was hidden by the technical start date from the admin UI template. This patch provide a fix for this issue. Also the sorting behavior for the event start and stop column is restored. In some cases the event start date wasn't set on ingest. Changing the duration of such an event in the event metadata tab in the admin UI has overridden the start time to the value of first januar 2001. This misbehavior is also adressed by this patch.