Smarter etc/ hints in documentation
This patches slightly changes the way tooltips for the configuration
directory are added to the docs. Instead of attaching it to every
occurrence of etc/, this patch only attaches it when the code starts with
etc/.

This is to avoid attaching this to blocks like `/etc/nginx/…` where we
already have an absolute path.