Make Thumbnail Optional
This patch makes the thumbnail generation in the video editor optional
since it causes major performance problems and/or failures on some
systems so that a number of Opencast adopters already ended up patching
or removing this in one way or another.

This does not change the default behavior of Opencast. With the default
configuration, this should make Opencast behave in the same way as
before.