Externalize & update Elasticsearch
This patch upgrades to the Elasticsearch High Level REST Client 7.4
which needs an Elasticsearch 7 node to work with.
Since embedding Elasticsearch is no longer supported, the embedded node
is removed. Elasticsearch 7 does no longer allow to store different
document types in the same index. Instead, one index per document type
is required. So the code was changed to create four "sub indices", one
for each document type. This means, that for the admin UI index, there
are actually 4 Elasticsearch indices now. The same applies for the
external API index.

So from now on, running an external Elasticsearch node is mandatory in
order to be able to run Opencast.

One small test case was removed as it would now require a running
external Elasticsearch node during execution time.