Make Opencast compatible with JDK11
This patch improves compatibility with newer JDKs. Namely, it is now
possible to build and run Opencast with JDK11. For JDK12 and JDK13,
building Opencast is possible as well, while running Opencast is not
possible with these JDKs, yet.

Since we want to stay compatible with JDK1.8 for now, language features
of newer JDKs cannot be used at the moment. Once we decide to drop
JDK1.8 support, new language features can be used without any further
preparations, except increasing the `jdk.version` in the main POM.

The following table shows the current state of Opencast compatibility
with some JDKs.

| JDK  | Build | Run | Language Features |
|------|-------|-----|-------------------|
|JDK1.8|  YES  | YES |        YES        |
|JDK11 |  YES  | YES |        NO         |
|JDK13 |  YES  | NO  |        NO         |

Compatibility with JDK 9, 10, and 12 was not tested, but for JDK 9 and
10, full (build _and_ runtime) compatibility is much likely given.

To also achieve runtime compatibility with JDK12 and JDK13, Karaf 4.3
is needed, which is not yet released. Once the new Karaf version is
available, enabling JDK12 and JDK13 runtime compatibility should be
fairly easy.

In order to achieve runtime compatibility with JDK11, a little quirk was
necessary: `jakarta.xml.bind-api-2.3.3.jar` had to be packaged with
Opencast as this version is required for JDK11. Version 2.3.3 of this
library is already released, but not yet uploaded to Maven Central
because it is part of the upcoming Java EE release, which was not yet
published. Once this is published, the packaged jar can most likely be
replaced with a Maven reference. Once Karaf comes with a Java EE version
including version 2.3.3 of jakarta.xml.bind-api or higher, it should no
longer be necessary to manually manage this dependency at all.

To enable compatibility with newer Java versions, lots of dependencies
were updated. Also, Java EE is currently in the process of renaming all
artifacts named `javax.something` to `jakarta.something-else` due to
legal issues with the term `javax`. Hence, a lot of pom files had to be
touched as upgrading a dependency also meant to rename it in many cases.

After upgrading all the dependencies, a few issues came up which are
worth being mentioned:

- JPA was complaining about duplicate column names in some queries. So I
  had to extend 3 `@JoinColumn` annotations.
- With Java 9, breaking changes to locales were made which caused some
  date formatting code parts to break. This was fixed by setting the
  JDK's `java.locale.providers` property to `COMPAT`.
- The `Lazy` class caused strange behavior where class files of classes
  using `Lazy` were compiled differently by different JDKs, even though
  both JDKs were compiling for the same byte version. This caused test
  failures. Since `Lazy` could be easily replaced and seemed to not
  provide any advantages, it was removed.
- Since Jetty had to be updated from version 1.x to version 2.x, some
  major changes to the `rest-test-environment` bundle were necessary.

The file `system.properties.append` used to configure XML processing
related factories to use internal JDK classes. This was changed to use
JDK-independent classes to improve compatibility with future JDKs.

Upgrading the ActiveMQ client library was not necessary, but was done
anyway.

Since a lot of dependencies were changed, the Maven part of `NOTICES`
was re-generated.