Editor Thumbnail Default
This patch changes the default value for enabling the thumbnail
generation in the video editor.  The new default is for the thumbnail
generation to be deactivated since it causes major performance problems
and/or failures on some systems so that a number of Opencast adopters
already ended up patching or removing this in one way or another.

The thumbnail generation was added in Opencast 6 and later made optional
due to ongoing issues in Opencast 7.x:

  https://github.com/opencast/opencast/pull/1220

Having this deactivated by default means that people need to
conscientiously activate this and will not accidentally run into
problems causing workflows to fail.