Send parameter as query for delete requests in rest docs
Form parameters for DELETE requests are currently not used and are not even
possible because we can't set them in the remote implementations in the
backend. However the rest docs currently use form parameters for everything
but GET, resulting in some endpoints with query parameters not working
correctly. This commit makes the rest docs send query parameters for
DELETE requests.