Fix Image Extraction From Short Videos
This image extraction for preview images was always done from a few
seconds after the beginning of the video to avoid extracting black
images. This causes the process and thus the workflow to completely fail
on very short videos as we have lately seen multiple times on the public
test servers.

Instead of failing, this patch will make Opencast extract the last frame
before the specified time instead. This behavior can still be changed
since it is only configuration but it's likely a much better default
configuration.

This fixes #1219