Video Segemntation On Short Videos
Our video segmentation works extremely bad on short videos, e.g. trying
to divide a 0.5 seconds video into 100 segments, causing the segments to
be shorter than a single frame which then caused problems with the
slide-preview extraction.

Assuming that a segmentation on extremely short videos is useless at
best and causes workflow failures or services in error states at worst,
this patch will make Opencast skip the segmentation on videos shorter
than 15 seconds.

This is related to #1219 and #1326