Drop X-Opencast-Matterhorn-Authorization
The `X-Opencast-Matterhorn-Authorization` header was used by Opencast to
identify that anonymous access should never be assumed, even if that
works according to the security specification to avoid ending up with
a wrong user internally.

This is necessary since HTTP Digest is a special case for Opencast
authentication and is handled separately from all other authentication.

But at the same time, HTTP Digest authentication also requires sending
the `X-Requested-Auth` header indicating that a particular authentication
type (HTTP Digest) is requested. This basically makes one of the two
headers redundant.

Worse, simply using X-Requested-Auth works in all cases where
authentication is required, making it seem as if just using this will
always work (which it mostly does) but there are some weird edge-cases
like calling `/info/me.json` which will return `anonymous` as a user.

To avoid this inconsistent behavior, this patch drops the header
`X-Opencast-Matterhorn-Authorization`. Sending it will not result in any
error, but the header will simply be ignored. This should ensure that
this change will not hurt any existing integrations using this header.