Dependency Tests
This patch fixes another bunch of dependency issues with modules and
introduces the maven-dependency-plugin to check for issues.

This additionally creates a file stating the current exceptions and adds
a test which is complaining if the actual state differs from these
stated exceptions. This will ensure that no new modules are added
without the dependency checks.